+++ src/AutotraderApi.php
@@ -16,9 +16,11 @@ use NorthBees\AutotraderApi\Exceptions\AutotraderWarning;
 use NorthBees\AutotraderApi\Traits\AutotraderAdvertisersTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderAuthenticationTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderCoDriverTrait;
+use NorthBees\AutotraderApi\Traits\AutotraderFinanceTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderFutureValuationsTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderHistoricValuationsTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderImageTrait;
+use NorthBees\AutotraderApi\Traits\AutotraderSearchTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderStockTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderTaxonomyTrait;
 use NorthBees\AutotraderApi\Traits\AutotraderValuationsTrait;
@@ -28,9 +30,11 @@ use NorthBees\AutotraderApi\Traits\AutotraderVehiclesTrait;
 class AutotraderApi
 {
     use AutotraderAuthenticationTrait;
+    use AutotraderFinanceTrait;
     use AutotraderFutureValuationsTrait;
     use AutotraderHistoricValuationsTrait;
     use AutotraderImageTrait;
+    use AutotraderSearchTrait;
     use AutotraderStockTrait;
     use AutotraderTaxonomyTrait;
     use AutotraderValuationsTrait;